change contractSize to double
git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@7166 19eef36c-86f0-de11-afab-0004ac968479

+++ common/src/main/java/ch/algotrader/entity/marketData/TickImpl.java
@@ -164,7 +164,7 @@ public class TickImpl extends Tick {
                 throw new IllegalStateException("SpreadSlope and SpreadConstant have to be defined to validate a tradeable security");
             }
 
-            int contractSize = family.getContractSize();
+            double contractSize = family.getContractSize();
             double maxSpreadSlope = family.getMaxSpreadSlope();
             double maxSpreadConstant = family.getMaxSpreadConstant();
