+++ common/src/main/java/ch/algotrader/entity/security/SecurityImpl.java
@@ -24,9 +24,6 @@ import org.apache.log4j.Logger;
 import org.hibernate.Hibernate;
 import org.springframework.beans.factory.annotation.Value;
 
-import com.espertech.esper.event.WrapperEventBean;
-import com.espertech.esper.event.bean.BeanEventBean;
-
 import ch.algotrader.ServiceLocator;
 import ch.algotrader.entity.marketData.MarketDataEvent;
 import ch.algotrader.entity.marketData.Tick;
@@ -35,6 +32,9 @@ import ch.algotrader.esper.EngineLocator;
 import ch.algotrader.util.MyLogger;
 import ch.algotrader.util.ObjectUtil;
 import ch.algotrader.util.metric.MetricsUtil;
+
+import com.espertech.esper.event.WrapperEventBean;
+import com.espertech.esper.event.bean.BeanEventBean;
 
 /**
  * @author <a href="mailto:aflury@algotrader.ch">Andy Flury</a>
@@ -123,7 +123,7 @@ public abstract class SecurityImpl extends Security {
         double marginPerContract = 0;
         if (marketDataEvent != null && marketDataEvent.getCurrentValueDouble() > 0.0) {
 
-            int contractSize = getSecurityFamily().getContractSize();
+            double contractSize = getSecurityFamily().getContractSize();
             marginPerContract = marketDataEvent.getCurrentValueDouble() * contractSize / initialMarginMarkup;
         } else {
             logger.warn("no last tick available or currentValue to low to set margin on " + this);
